Make object nodes easier to use
Before this change ObjectNode had it's children in an array and not a HashMap which meant that sometimes it was hard to get certain key/value pairs out of the object (had to create a new HashMap of all the key/value pairs). Now I just do that when creating the ObjectNode.

If a save file object has duplicate keys in the same object then the ObjectNode changes to PropertyNode that has the same key and a ListNode of children.