Change GameFile List<PropertyNode> to HashMap<String, Object>
If the save file had duplicate keys on the "first layer" the duplicate keys would be saved to the property node list and not added under a their own list. That used to mean that not everything was printed out when creating the JSON save file because JSON doesn't allow duplicate keys.