Lower memory usage by saving big objects as strings
Big objects like countries don't need to be saved as ObjectNodes, they can just be saved as a string to save memory. This will slow down parsing but about the same time would have been used when converting the ObjectNode to a JSON string when writing to file.

This should be only used for objects that have unique keys because changing a JSON object string to a JSON array will be awful.